+++ packages/scaffold/src/client.ts
@@ -25,10 +25,12 @@ import {
   PublicStateController,
   ThemeController,
   SnackController,
-  RouterController
+  RouterController,
+  EnsController
 } from '@web3modal/core'
 import { setColorTheme, setThemeVariables } from '@web3modal/ui'
 import type { SIWEControllerClient } from '@web3modal/siwe'
+import { ConstantsUtil } from '@web3modal/common'
 
 // -- Helpers -------------------------------------------------------------------
 let isInitialized = false
@@ -144,6 +146,10 @@ export class Web3ModalScaffold {
   }
 
   // -- Protected ----------------------------------------------------------------
+  protected replace(route: RouterControllerState['view']) {
+    RouterController.replace(route)
+  }
+
   protected redirect(route: RouterControllerState['view']) {
     RouterController.push(route)
   }
@@ -158,6 +164,12 @@ export class Web3ModalScaffold {
 
   protected isTransactionStackEmpty() {
     return RouterController.state.transactionStack.length === 0
+  }
+
+  protected isTransactionShouldReplaceView() {
+    return RouterController.state.transactionStack[
+      RouterController.state.transactionStack.length - 1
+    ]?.replace
   }
 
   protected setIsConnected: (typeof AccountController)['setIsConnected'] = isConnected => {
@@ -247,6 +259,18 @@ export class Web3ModalScaffold {
       AccountController.setPreferredAccountType(preferredAccountType)
     }
 
+  protected getWalletConnectName: (typeof EnsController)['getNamesForAddress'] = address => {
+    return EnsController.getNamesForAddress(address)
+  }
+
+  protected resolveWalletConnectName = async (name: string) => {
+    const trimmedName = name.replace(ConstantsUtil.WC_NAME_SUFFIX, '')
+    const wcNameAddress = await EnsController.resolveName(trimmedName)
+    const networkNameAddresses = Object.values(wcNameAddress?.addresses) || []
+
+    return networkNameAddresses[0]?.address || false
+  }
+
   // -- Private ------------------------------------------------------------------
   private async initControllers(options: ScaffoldOptions) {
     NetworkController.setClient(options.networkControllerClient)
@@ -263,12 +287,6 @@ export class Web3ModalScaffold {
     OptionsController.setCustomWallets(options.customWallets)
     OptionsController.setEnableAnalytics(options.enableAnalytics)
     OptionsController.setSdkVersion(options._sdkVersion)
-
-    if (options.siweControllerClient) {
-      const { SIWEController } = await import('@web3modal/siwe')
-
-      SIWEController.setSIWEClient(options.siweControllerClient)
-    }
 
     if (options.metadata) {
       OptionsController.setMetadata(options.metadata)
@@ -293,6 +311,13 @@ export class Web3ModalScaffold {
     if (options.allowUnsupportedChain) {
       NetworkController.setAllowUnsupportedChain(options.allowUnsupportedChain)
     }
+
+    if (options.siweControllerClient) {
+      const { SIWEController } = await import('@web3modal/siwe')
+
+      SIWEController.setSIWEClient(options.siweControllerClient)
+    }
+
     ConnectionController.setClient(options.connectionControllerClient)
   }