+++ packages/scaffold/src/partials/w3m-account-tokens-widget/styles.ts
@@ -14,4 +14,8 @@ export default css`
     overflow: scroll;
     scrollbar-width: none;
   }
+
+  .contentContainer::-webkit-scrollbar {
+    display: none;
+  }
 `